+++ lib/oxm/sax_handler.rb
@@ -1,6 +1,8 @@
-module OXM
+class OXM
   class SaxHandler < Nokogiri::XML::SAX::Document
     attr_reader :outputs
+    attr_reader :warnings
+    attr_reader :errors
 
     def initialize loop_node, &block
       @loop_node = loop_node
@@ -8,6 +10,17 @@ module OXM
       @tags = []
       @objects = []
       @outputs = []
+
+      @warnings = []
+      @errors = []
+    end
+
+    def warning str
+      @warnings << str
+    end
+
+    def error str
+      @errors << str
     end
 
     def start_element tag, attributes
@@ -34,8 +47,17 @@ module OXM
 
     def end_element tag
       unless @objects.empty?
+        obj = @objects.pop
+
+        # Strip content
+        if obj.cdata?
+          obj.cdata = obj.to_s.strip
+        else
+          obj.content = obj.to_s.strip
+        end
+        obj.content = nil if obj.to_s.empty?
+
         if match?
-          obj = @objects.pop
           if @block
             @block.call obj
           else
@@ -43,7 +65,6 @@ module OXM
           end
           @object = nil
         else
-          @objects.pop
           @object = @objects.last
         end
       end